feat: Add contact, settings, dashboard, privacy policy, and terms of service pages
Add new pages for contact, settings, dashboard, privacy policy, and terms of service. These pages provide additional functionality and information to the users.

auth final

Finalize the authentication feature. All necessary changes and improvements have been made to ensure a secure and smooth authentication process.

first commit

Initial commit of the project. This commit includes the basic project structure and configuration files.

Update migration_lock.toml

Update the migration_lock.toml file to specify the provider as "postgresql".

Update package.json

Add "@types/jest" as a dev dependency in the package.json file.

Update schema.prisma

Change the type of "endDate" field in the schema.prisma file from DateTime to String.

Add TodoType interface

Create a new interface called TodoType in the TodoType.ts file. This interface defines the structure of a todo object.

Update dashboard page

Update the dashboard page to include an AddTodoButton and a TodoContainer component. These components enhance the functionality and user experience of the dashboard.

Add GetTodos action

Implement the GetTodos action in the GetTodos.ts file. This action retrieves the todos associated with the logged-in user.

Add Textarea component

Create a new Textarea component in the textarea.tsx file. This component provides a customizable textarea input for forms.

Add AddTodoFormSchema

Create a new schema called AddTodoFormSchema in the AddTodoFormSchema.ts file. This schema defines the validation rules for the fields in the add todo form.

Update TodoContainer component

Update the TodoContainer component to fetch and display the todos associated with the logged-in user.

Add Popover component

Create a new Popover component in the popover.tsx file. This component provides a customizable popover UI element.

Update Todo component

Update the Todo component to display the details of a todo, including its title, description, status, completion indicator, created date, and end date.

Add AddTodo action

Implement the AddTodo action in the AddTodo.ts file. This action adds a new todo to the database and associates it with the logged-in user.